+++ tests/unit/test_application.py
@@ -16,6 +16,7 @@
 """Unit tests for application classes."""
 import json
 import os
+import sys
 from textwrap import dedent
 
 import craft_cli
@@ -27,9 +28,9 @@ from craft_parts.packages import snaps
 from craft_providers import bases
 
 from snapcraft import application, services
-from snapcraft.errors import ClassicFallback
+from snapcraft.errors import ClassicFallback, SnapcraftError
 from snapcraft.models.project import Architecture
-from snapcraft.parts.yaml_utils import CURRENT_BASES, ESM_BASES
+from snapcraft.parts.yaml_utils import CURRENT_BASES, ESM_BASES, LEGACY_BASES
 
 
 @pytest.fixture(
@@ -42,6 +43,32 @@ from snapcraft.parts.yaml_utils import CURRENT_BASES, ESM_BASES
 )
 def architectures(request):
     return request.param
+
+
+@pytest.fixture()
+def mock_confirm(mocker):
+    return mocker.patch(
+        "snapcraft.commands.remote.confirm_with_user", return_value=True
+    )
+
+
+@pytest.fixture()
+def mock_remote_build_run(mocker):
+    _mock_remote_build_run = mocker.patch(
+        "snapcraft.commands.remote.RemoteBuildCommand._run"
+    )
+    return _mock_remote_build_run
+
+
+@pytest.fixture()
+def mock_run_legacy(mocker):
+    return mocker.patch("snapcraft_legacy.cli.legacy.legacy_run")
+
+
+@pytest.fixture()
+def mock_remote_build_argv(mocker):
+    """Mock `snapcraft remote-build` cli."""
+    return mocker.patch.object(sys, "argv", ["snapcraft", "remote-build"])
 
 
 @pytest.mark.parametrize("env_vars", application.MAPPED_ENV_VARS.items())
@@ -331,3 +358,113 @@ def test_esm_pass(mocker, snapcraft_yaml, base):
         pass
     else:
         mock_dispatch.assert_called_once()
+
+
+@pytest.mark.parametrize(
+    "envvar", ["force-fallback", "disable-fallback", "badvalue", None]
+)
+@pytest.mark.parametrize("base", CURRENT_BASES - {"core22"})
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar(
+    monkeypatch,
+    snapcraft_yaml,
+    base,
+    envvar,
+    mock_remote_build_run,
+    mock_run_legacy,
+):
+    """Bases core24 and later run new remote-build regardless of envvar."""
+    snapcraft_yaml_dict = {"base": base, "build-base": "devel", "grade": "devel"}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+
+    if envvar:
+        monkeypatch.setenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", envvar)
+    else:
+        monkeypatch.delenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", raising=False)
+
+    application.main()
+
+    mock_remote_build_run.assert_called_once()
+    mock_run_legacy.assert_not_called()
+
+
+@pytest.mark.parametrize("base", LEGACY_BASES)
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar_disable_fallback_core20(
+    snapcraft_yaml, base, mock_remote_build_run, mock_run_legacy, monkeypatch
+):
+    """core20 base run new remote-build if envvar is `disable-fallback`."""
+    monkeypatch.setenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", "disable-fallback")
+    snapcraft_yaml_dict = {"base": base}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+
+    application.main()
+
+    mock_remote_build_run.assert_called_once()
+    mock_run_legacy.assert_not_called()
+
+
+@pytest.mark.parametrize("base", LEGACY_BASES | {"core22"})
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar_force_fallback_core22(
+    snapcraft_yaml, base, mock_remote_build_run, mock_run_legacy, monkeypatch
+):
+    """core22 and older bases run legacy remote-build if envvar is `force-fallback`."""
+    monkeypatch.setenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", "force-fallback")
+
+    snapcraft_yaml_dict = {"base": base}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+    application.main()
+
+    mock_run_legacy.assert_called_once()
+    mock_remote_build_run.assert_not_called()
+
+
+@pytest.mark.parametrize("base", LEGACY_BASES)
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar_force_fallback_unset_core20(
+    snapcraft_yaml, base, mock_remote_build_run, mock_run_legacy, monkeypatch
+):
+    """core20 base run legacy remote-build if envvar is unset."""
+    monkeypatch.delenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", raising=False)
+
+    snapcraft_yaml_dict = {"base": base}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+    application.main()
+
+    mock_run_legacy.assert_called_once()
+    mock_remote_build_run.assert_not_called()
+
+
+@pytest.mark.parametrize("base", {"core22"})
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar_force_fallback_empty_core22(
+    snapcraft_yaml, base, mock_remote_build_run, mock_run_legacy, monkeypatch
+):
+    """core22 bases run craft-application remote-build if envvar is empty."""
+    monkeypatch.setenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", "")
+
+    snapcraft_yaml_dict = {"base": base}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+    application.main()
+
+    mock_remote_build_run.assert_called_once()
+    mock_run_legacy.assert_not_called()
+
+
+@pytest.mark.parametrize("base", LEGACY_BASES | {"core22"})
+@pytest.mark.usefixtures("mock_confirm", "mock_remote_build_argv")
+def test_run_envvar_invalid(snapcraft_yaml, base, monkeypatch):
+    """core20 and core22 bases raise an error if the envvar is invalid."""
+    monkeypatch.setenv("SNAPCRAFT_REMOTE_BUILD_STRATEGY", "badvalue")
+
+    snapcraft_yaml_dict = {"base": base}
+    snapcraft_yaml(**snapcraft_yaml_dict)
+    with pytest.raises(SnapcraftError) as err:
+        application.main()
+
+    assert err.match(
+        "Unknown value 'badvalue' in environment variable "
+        "'SNAPCRAFT_REMOTE_BUILD_STRATEGY'. Valid values are 'disable-fallback' and "
+        "'force-fallback'"
+    )
